Extract DeleteNextCharActionTest
